Refactored events into one sub-module per event
